Add parameter to limit the animation frame rate
Using the new --animation-fps parameter, users can now limit
the maximum number of frames per second that they will get from
the data source. Note that this won't affect the rendering rate,
but just the animation rate. By default it gets a value of 0,
setting a new frame as soon as the previous one is finished